backend/requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
pydantic>=2.0
python-multipart>=0.0.6
pypdf>=4.0
python-docx>=1.0
pytest>=7.0
